+++ src/article/article.service.ts
@@ -12,7 +12,7 @@ import { UpdateArticleRequestDto } from './dto/request/update-article-request.dt
 import { Article } from './entities/article.entity';
 import { CategoryService } from '@root/category/category.service';
 import { FindAllBestDto } from '@root/best/dto/find-all-best.dto';
-import { PageDto } from '@root/pagination/pagination.dto';
+import { PageDto } from '@root/pagination/dto/pagination.dto';
 import { ReactionService } from '@root/reaction/reaction.service';
 import { Category } from '@root/category/entities/category.entity';
 import { User } from '@root/user/entities/user.entity';
