+++ src/article/dto/request/find-article-request.dto.ts
@@ -1,6 +1,6 @@
 import { IsInt, IsNotEmpty, IsOptional, Min } from 'class-validator';
 import { ApiPropertyOptional } from '@nestjs/swagger';
-import { PageOptionsDto } from '@root/pagination/page-options.dto';
+import { PageOptionsDto } from '@root/pagination/dto/page-options.dto';
 import { Type } from 'class-transformer';
 
 export class FindArticleRequestDto extends PageOptionsDto {
