+++ src/pagination/interface/pagination.enum.ts
